+++ spec/unit/resource/api500/synergy/server_hardware_type_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::API500::Synergy::ServerHardwareType do
+  include_context 'shared context'
+
+  it 'inherits from OneviewSDK::API300::Synergy::ServerHardwareType' do
+    expect(described_class).to be < OneviewSDK::API300::Synergy::ServerHardwareType
+  end
+end