+++ CHANGELOG.md
@@ -19,6 +19,7 @@ This release adds support to OneView Rest API version 500 for the hardware varia
    - SAS Logical Interconnect
    - SAS Logical Interconnect Group
    - Scopes
+   - Server Hardware Type
    - Switches
 
 #### Breaking changes